+++ examples/unbalanced-partial/plot_partial_wass_and_gromov.py
@@ -125,10 +125,10 @@ C2 = sp.spatial.distance.cdist(xt, xt)
 # transport 100% of the mass
 print('------m = 1')
 m = 1
-res0, log0 = ot.partial.partial_gromov_wasserstein(C1, C2, p, q, m=m, log=True)
-res, log = ot.partial.entropic_partial_gromov_wasserstein(C1, C2, p, q, 10,
-                                                          m=m, log=True,
-                                                          verbose=True)
+res0, log0 = ot.gromov.partial_gromov_wasserstein(C1, C2, p, q, m=m, log=True)
+res, log = ot.gromov.entropic_partial_gromov_wasserstein(C1, C2, p, q, 10,
+                                                         m=m, log=True,
+                                                         verbose=True)
 
 print('Wasserstein distance (m = 1): ' + str(log0['partial_gw_dist']))
 print('Entropic Wasserstein distance (m = 1): ' + str(log['partial_gw_dist']))
@@ -146,11 +146,11 @@ pl.show()
 # transport 2/3 of the mass
 print('------m = 2/3')
 m = 2 / 3
-res0, log0 = ot.partial.partial_gromov_wasserstein(C1, C2, p, q, m=m, log=True,
-                                                   verbose=True)
-res, log = ot.partial.entropic_partial_gromov_wasserstein(C1, C2, p, q, 10,
-                                                          m=m, log=True,
-                                                          verbose=True)
+res0, log0 = ot.gromov.partial_gromov_wasserstein(C1, C2, p, q, m=m, log=True,
+                                                  verbose=True)
+res, log = ot.gromov.entropic_partial_gromov_wasserstein(C1, C2, p, q, 10,
+                                                         m=m, log=True,
+                                                         verbose=True)
 
 print('Partial Wasserstein distance (m = 2/3): ' +
       str(log0['partial_gw_dist']))